refactor: extract Authly web apps into lib/authly-web
