+++ lib/authly-domain/src/repo/mod.rs
@@ -1,2 +1,4 @@
+pub mod crypto_repo;
 pub mod entity_repo;
+pub mod object_repo;
 pub mod session_repo;
