+++ src/db/mod.rs
@@ -4,7 +4,6 @@ pub mod directory_db;
 pub mod document_db;
 pub mod init_db;
 pub mod oauth_db;
-pub mod object_db;
 pub mod policy_db;
 pub mod service_db;
 pub mod settings_db;
